fz.qlwz_letv8key_switch, fz.javis_lock_report, fz.sihas_people_cnt  and fz.sihas_action
